requirements.txt
flask
faster-whisper
gunicorn
openai-whisper